Renamed LALRPOP-generated parsing methods.
